Implement base test case class to avoid repeated logic
The base test case class provides the
following:
- A read file function that uses the correct
  encoding (#18).
- A variable to set all the difference amount
  for all tests at once.

Classes using this base class need to call the
parent's initialiser and also set the tag_name
variable. This is used to identify the folder
of test case files.

The base test case class could also be set to
be abstract, to prevent a plain base test case
being created.